folder updation+forge fmt

+++ contracts/PushCore/PushCoreV3.sol
@@ -918,11 +918,11 @@ contract PushCoreV3 is
 
     function migrateAddresToBytes32(address[] calldata _channels) external whenPaused {
         onlyPushChannelAdmin();
-        for(uint i; i < _channels.length; ++i){
-           CoreTypes.Channel memory _channelData = channels[_channels[i]];
+        for (uint256 i; i < _channels.length; ++i) {
+            CoreTypes.Channel memory _channelData = channels[_channels[i]];
             bytes32 _channelBytesID = BaseHelper.addressToBytes32(_channels[i]);
             channelInfo[_channelBytesID] = _channelData;
             delete channels[_channels[i]];
         }
-    } 
+    }
 }
